requirements.txt
google-cloud-pubsub==2.4.1
tweepy==3.10.0